x86: add support for marking a device memory region at boot
The ARM boot protocol lets the bootloader pass an additional region of
memory that will be marked as device memory by the kernel. This is
used by sel4cp to prevent memory regions already populated by the
loader from being reused. There are other use cases.

This commit enables a similar feature on x86. An artificial multiboot2
boot info tag is used to pass the base address and size of the
additional region. A compliant multiboot boot loader is necessary to
pass the additional tag, such as the loader from sel4cp/x86. Note that
this change is transparent if the tag is not exposed by the loader.